Make error handling more principled and future proof
We use thiserror to define proper errors types for errors that are
exposed in the API.

We also annotate the errors as #[non_exhaustive] to make sure that we
can extend them in the future.

For the example we use anyhow.

Background info: failure was deprecated and anyhow and thiserror are
suggested as replacements going forward.

Fixes #6